mail-filter/amavisd-new: new revision sans the app-arch/freeze dependency.
The app-arch/freeze package fails to build with a modern compiler and
glibc. It's probably time to lay this package to rest. The last
upstream release is from circa 1999. Fedora has some patches against
it, but an email from the author (in the Fedora RPM) suggests that
even he thinks the package is dead.

While it's fun to see a long list of supported file types in amavisd,
I've never actually seen one of these archives in real life. The
recent build failures are due to some code quality issues; so even if
it's possible to patch them, one has to ask: do we want to be feeding
this unmaintained program untrusted input from the network? With all
things considered, its removal is for the best.

Bug: https://bugs.gentoo.org/669206
Signed-off-by: Michael Orlitzky <mjo@gentoo.org>
Package-Manager: Portage-2.3.51, Repoman-2.3.11