media-libs/vidstab: Video stabilization library
This is a second attempt for PR, it was discussed in
https://github.com/gentoo/gentoo/pull/8250 before.

This is a standalone library used by ffmpeg, so submitting it first
here, once in the tree, I'll submit PR to add it (compile time) to
ffmpeg  (currently tested fine with ffmpeg-{3.3.6,4.1}).

If anyone can hint me in the right direction to enable the tests in
src_test, I'd do the rest. Basically there is one unit test (13 checks)
in tests/ and to manually run them, one needs to run from ${S}:

cd tests/cmake
cmake ../
make
./tests

But after fiddling with cmake-utils_src_test I couldn't figure it out...

Thanks @candrews for helping me through the process.

Closes: https://bugs.gentoo.org/517892
Signed-off-by: Kalin KOZHUHAROV <kalin@thinrope.net>