gnome-extra/gnome-shell-extensions-topicons-plus: bump to v22 for gnome-shell-3.30 compat
This extension is in practice not maintained upstream anymore, and is at a
technical dead end for wayland (can't work). Some alternatives to research
might be:
https://extensions.gnome.org/extension/615/appindicator-support/
https://extensions.gnome.org/extension/1497/topicons-redux/ (as a maintained
version of X11 topicons)

Meanwhile lets have an upgraded version that at least will continue to work on
GNOME 3.30 X11 sessions.

Package-Manager: Portage-2.3.52, Repoman-2.3.12
Signed-off-by: Mart Raudsepp <leio@gentoo.org>