gnome-base/gnome-3.24.2: allow gnome-core-apps-3.26 upgrade
gnome-core-apps dropped IUSE=cdr due to brasero being rather unmaintained.
Allow installing gnome-3.24[cdr] with gnome-core-apps-3.26 by assuming it
exists, even if it doesn't (instead of not accepting the upgrade), as this
is a metapackage and it should be fine that the IUSE=cdr doesn't do anything
during migration to 3.26.

Package-Manager: Portage-2.3.52, Repoman-2.3.12
Signed-off-by: Mart Raudsepp <leio@gentoo.org>