net-analyzer/nagios-core: new revision to fix CGI install path.
Upstream commit 37da6f39c8 fixes an old bug in the build system, where
CGIDIR was set to @sbindir@ (rather than to @cgibindir@), and HTMLDIR
was set to @datarootdir@ (rather than to @webdir@). Now that the bug
is fixed, the ebuild has been revisioned to use the right configure flags:
we pass --with-cgibindir the old argument of --sbindir, and --with-webdir
the old argument of --datadir.

Thanks to Tomáš Mózes for the report and the fix.

Closes: https://bugs.gentoo.org/675872
Signed-off-by: Michael Orlitzky <mjo@gentoo.org>
Package-Manager: Portage-2.3.51, Repoman-2.3.11