sys-apps/groff: Fix cross-compiling following build file changes
Override the variables when calling make instead of changing the
Makefiles. The comment in the Makefiles suggests that this is how
upstream wants you to do it.

Using ${EPREFIX} as before is not correct as this refer to the target
prefix, not the build prefix. EAPI 7 provides ${BROOT} for this but
for tools in the path, it is better to use type -P instead.

groff is not in @system so there is no guarantee that it will actually
be present on the build system. EAPI 7 did not include EAPI
5-hdepend's targetroot USE flag feature, which prevents us from
putting groff in BDEPEND. The preferred approach is to do a native
build in tandem with a cross build but this makes the ebuild more
complex so I will attempt this in a separate fix.

Closes: https://bugs.gentoo.org/674866
Package-Manager: Portage-2.3.58, Repoman-2.3.12
Signed-off-by: James Le Cuirot <chewi@gentoo.org>