dev-java/openjdk: bump to 11.0.2+7
make cups optional. it's still required to build, but can be
removed later.

Removed jvm variants support. It was incomplete and overly complex.
Greater flexibility can be achieved via features.

If you want to enable or disable certain features you can pass
--with-jvm-features=[-feature|feature] to ebuild via EXTRA_ECONF
The following JVM features are available:

  compiler1 compiler2 zero minimal dtrace jvmti jvmci
  graal vm-structs jni-check services management cmsgc
  epsilongc g1gc parallelgc serialgc zgc nmt cds
  static-build link-time-opt aot jfr

Which are valid to use depends on the target platform

Note: in this version zgc (zero garbage collector) is not enabled yet.
It will be enabled by default on supported arches starting with 11.0.3
https://wiki.openjdk.java.net/display/zgc/Main

to enable it for this build pass --with-jvm-features=zgc via EXTRA_ECONF

Package-Manager: Portage-2.3.58, Repoman-2.3.12
Signed-off-by: Georgy Yakovlev <gyakovlev@gentoo.org>