kde5.eclass: Relocate KDE_INSTALL_DOCBUNDLEDIR to /usr/share/help
These are handbook files installed by hundreds of packages to a
central location, to be accessed by KDE Helpcenter application,
with compressed index cache files.

This leads to the following QA issues on every package w/ USE handbook:
* The ebuild is installing to one or more unexpected paths
* One or more compressed files were found in docompress-ed directories

We'll use the EAPI-7 transition to move the location from previously
/usr/share/doc/HTML to /usr/share/help [1]. Transition for khelpcenter
will be seemless as kde-frameworks/kdoctools can be easily patched to
accept both locations, until eventually /usr/share/doc/HTML is empty.

As a side-effect, we will be able to drop the workaround for disabling
auto compression of files beneath /usr/share/doc together with EAPI-6.

[1] See also: https://www.freedesktop.org/wiki/Specifications/help-system/
    It is also the path upstream intends to migrate to in the future.

Bug: https://bugs.gentoo.org/667138
Signed-off-by: Andreas Sturmlechner <asturm@gentoo.org>