net-analyzer/gvm-tools: New package.
gvm-tools in short is a collection of tools that help
with remote controlling a Greenbone Security Manager (GSM)
appliance and its underlying Greenbone Vulnerability Manager (GVM).
The tools essentially aid accessing the communication protocols
GMP (Greenbone Management Protocol) and OSP (Open Scanner Protocol).
This package is part of net-analyzer/openvas.
This package is replacement of net-analyzer/openvas-cli
which is no needed anymore.
This package is replacement of net-analyzer/openvas-tools
which is no needed anymore.

Reported-by: NP-Hardass <np-hardass@gentoo.org>
Acked-by: Hasan ÇALIŞIR <hasan.calisir@psauxit.com>
Tested-by: Hasan ÇALIŞIR <hasan.calisir@psauxit.com>
Signed-off-by: Hasan ÇALIŞIR <hasan.calisir@psauxit.com>
Package-Manager: Portage-2.3.51, Repoman-2.3.11