x11-libs/gtk+: Further fallback fixes for GtkSwitch unicode glyphs
Add a final fallback to no glyphs at all, so if the glyphs aren't
found, it'll at least just not use any and not mess up layout and
stomp on some memory, etc. The ON or OFF position should still be
visible from color and position, at least, but for most people at
least the secondary glyphs should be available.
gtk+-3.22.5 changes all this to icons, but it also changes the
default theme look, so we do this simpler patch for 3.22.4 for
that to be the next stable candidate soon, instead of theme
changing 3.22.5.

Closes: https://bugs.gentoo.org/676098
Package-Manager: Portage-2.3.52, Repoman-2.3.12
Signed-off-by: Mart Raudsepp <leio@gentoo.org>