media-libs/grilo: bump to 0.3.7
USE=examples is dropped, because the C files are shown as
examples in gtk-doc anyways with USE=gtk-doc.
liboauth is now properly a dep only for the gtk test-ui by
upstream, so we can follow.
Python tests aren't used for now, so tests don't need
introspection and specific python stuff anymore; however
meson build does not require python for some code generation.
The SLOTting is ported to meson as well.

Package-Manager: Portage-2.3.52, Repoman-2.3.12
Signed-off-by: Mart Raudsepp <leio@gentoo.org>