app-text/diff-pdf: new package to visually compare two PDF files.
This has been an open request (and desire of mine) for a while, but
the older releases depended on the deprecated 2.8 slot of wxGTK.
However a new release was recently made with all of the upstream
improvements from the past few years, so it is now possible to
package this properly. Here it is.

Thanks to Erik Quaeghebeur who opened the package request bug,
supplied an ebuild, and helped me do some testing.

Closes: https://bugs.gentoo.org/549260
Reported-by: Erik Quaeghebeur
Signed-off-by: Michael Orlitzky <mjo@gentoo.org>
Package-Manager: Portage-2.3.51, Repoman-2.3.11