games-board/gnome-mines: depend on librsvg[vala]
vala itself doesn't ship vapi files for librsvg since 0.40,
so we need to depend on librsvg[vala] for them to be provided.
We do this unconditionally even if older vala is used for build,
as they are more up to date and matching librsvg C library than
what vala itself shipped.

Closes: https://bugs.gentoo.org/678212
Package-Manager: Portage-2.3.52, Repoman-2.3.12
Signed-off-by: Mart Raudsepp <leio@gentoo.org>