package.mask: Temporarily mask gnome-desktop-3.30
It breaks ABI, so need to see what doesn't work with it, and also
upgrade some of the consumers at the same time to save on build
cycles for users. It will be an actively changing mask block with
items gradually added for a set, then all unmasked together and
then another new set of packages being worked on in mask, etc.

Signed-off-by: Mart Raudsepp <leio@gentoo.org>