gnome-base/gdm: bump to 3.30.3 - now always runs on VT1
GDM 3.30 has support for killing the login screen when not in use
and the login screen (gdm used gnome-shell) is running with wayland.
This will save a lot of memory when effective, as gnome-shell and
I believe also gnome-settings-daemon and its plugins don't need to
keep running anymore after having logged in. Memory wins between
100-300MB are to be expected.
However if wayland isn't used or gnome-shell crashes with wayland
(probably quite likely with gnome-shell-3.26 or older), then the
only win is not having an agetty opened on VT1, which saves some
~600kB RAM.

All this new code assumes the default VT1 is used for gdm. If we
keep passing VT7 for initial VT, gdm just never shows a login screen
until user manually switches to VT7 with Alt+F7. Instead of making
that work, just always use VT1 for now like a good modern distro.
We will see later how this works out for non-systemd, but currently
the package still hard requires systemd anyways.

Also drops obsolete sessreg dependency - it was used in the PostSession
and co script in the past, but doesn't seem to have been for a long
long time. If someone still configures it to be called in their
local session scripts, they'll have to install it themselves, as we
don't need it imposed on everyone when it's not used out of the box.

Package-Manager: Portage-2.3.52, Repoman-2.3.12
Signed-off-by: Mart Raudsepp <leio@gentoo.org>