gnome-base/gdm: require older gnome-settings-daemon for older gdm
org.gnome.SettingsDaemon.A11yKeyboard is a required component in
<gdm-3.27.90 versions, which won't exist in gnome-settings-daemon-3.28
anymore. So for older gdm require <g-s-d-3.27.
Newer gdm is mostly fine with old gnome-settings-daemon, as it just
doesn't launch that component anymore, which means onscreen keyboard
doesn't work anymore until g-s-d is upgraded - but this is ~arch for
now and the newer gdm and g-s-d versions will be stabled together for
sure in the future (unless the timed login security fixes are deemed
too important).

Package-Manager: Portage-2.3.52, Repoman-2.3.12
Signed-off-by: Mart Raudsepp <leio@gentoo.org>