net-analyzer/monitoring-plugins: make 'check_ping -4' call 'ping -4'
Currently, when calling 'check_ping -4' on a host that has both A and AAAA
records, the AAAA record is used. This is because the IPv4 ping command
is set to 'ping' at configure-time, when it should be 'ping -4'.

For consistency, this commit also changes the IPv6 ping command from
'ping6' to 'ping -6'.

The upstream issue is
https://github.com/monitoring-plugins/monitoring-plugins/issues/898.
Changing the ping command at configure-time was proposed in
https://github.com/monitoring-plugins/monitoring-plugins/pull/1531.

Package-Manager: Portage-2.3.62, Repoman-2.3.12
Signed-off-by: Louis Sautier <sbraz@gentoo.org>
Closes: https://github.com/gentoo/gentoo/pull/11169