app-misc/tracker-miners: bump to 2.1.6, workaround gst sandbox issues
Disables guarantee_metadata option by default, otherwise mainly just
most of the patches we needed for 2.1.5 are included now, plus
translation updates.
Workaround gstreamer sandbox issues by avoiding the gst-inspect
checking and patching the detection result to true or false, depending
on if gstreamer is used or not. Theoretically the test should succeed
always with ffmpeg, but not bothering to figure out if there is some
USE flags on ffmpeg needed to ensure that or not, etc, so just skipping
them with ffmpeg for the time being.

Closes: https://bugs.gentoo.org/678442
Package-Manager: Portage-2.3.52, Repoman-2.3.12
Signed-off-by: Mart Raudsepp <leio@gentoo.org>