gnome-base/gnome-control-center: fix runtime crash on wayland from build fix
Unfortunately we will need to lock USE=wayland between gtk+ and here for now.
If gtk+ supports wayland, it will run gnome-control-center with native
wayland, and when g-c-c doesn't have the code to match, it'll try to iterate
devices with X11 methods, which will crash just like it did on full wayland
(USE=wayland on both) before the fix in this commit.
One option could be to restrict gdk allowed backends to not include wayland
if wayland support isn't included, but that's something to tackle on top of
3.32 in the future.

Closes: https://bugs.gentoo.org/679042
Package-Manager: Portage-2.3.52, Repoman-2.3.12
Signed-off-by: Mart Raudsepp <leio@gentoo.org>