x11-wm/mutter: require gles2 for wayland, try to fix non-gles2 builds
Native backend makes use of gles3 for some features, and I'm not sure if
these gracefully degrade or not without gles3. So require gles2 with
wayland (which is when we  enable build of native backend), which also
allows for a simpler hacky patch for fixing non-gles2 builds when system
doesn't have USE=gles2 on mesa.
I consider this quite ugly, but
1) this is ~arch
2) this will be better in the meson port in mutter-3.32
3) it hopefully works well enough anyways, sans potentially unnecessary
mesa[gles2] dep and full gles2 cogl builds.
People who mind the gles2 requirement are welcome to come up with
something better. It should somehow express the feature dependency of
this, even if just metadata.xml descriptions.

Closes: https://bugs.gentoo.org/679074
Package-Manager: Portage-2.3.52, Repoman-2.3.12
Signed-off-by: Mart Raudsepp <leio@gentoo.org>