gnome-base/gnome-shell: add explicit Cantarell font dep
gnome-shell uses Cantarell as its global font-family CSS
setting, so pull it in explicitly. This is also the main
font used by GNOME overall for its design looks, etc, and
we removed the explicit dep from lower of the stack. Thus
it's important to have it somewhere like gnome-shell for
that reason as well.

Package-Manager: Portage-2.3.52, Repoman-2.3.12
Signed-off-by: Mart Raudsepp <leio@gentoo.org>