gnome-base/nautilus: bump to 3.28.1
nautilus-desktop is now gone. Suggested alternative: "Desktop Icons"
from https://extensions.gnome.org

Tracker support is not optional anymore. It is theoretically possible
to patch it, but very cumbersome and hard to maintain on bumps - it
uses tracker for new "Starred" files feature and that is intertwined
a lot more than batch rename and tracker search were.
However, if one doesn't care for "tracker search" bits, one could avoid
tracker-miners and only have the core - the miners are not used for
batch rename and (probably) starring, as tracker provides more than
mining metadata from your files, and this is what nautilus relies on
to not NIH it.

nautilus-sendto extension code also not optional anymore, as it's a
rather tiny thing and better to have it always built, so no need to
rebuild all of nautilus after deciding to install nautilus-sendto for
its feature. It will probably go away with GNOME 3.34 in favor of
desktop portals for mail sending, etc, so it would work in flatpak too.

XMP and EXIF support isn't gone, it's just done via gexiv2 now.

Package-Manager: Portage-2.3.52, Repoman-2.3.12
Signed-off-by: Mart Raudsepp <leio@gentoo.org>