net-libs/libsoup: default enable USE=vala
Various GNOME packages will unconditionally need it. Default enable
it, as it just means need of a vala slot at build time. Vala itself
ships a libsoup vapi too, but it should be better to have it from
the library itself, so it matches the API the library version
actually provides.

Bug: https://bugs.gentoo.org/665578
Package-Manager: Portage-2.3.52, Repoman-2.3.12
Signed-off-by: Mart Raudsepp <leio@gentoo.org>