www-apps/radicale: fix file permissions for state directory
The "diropts" was not effective and the folder might have been created
with incorrect owner and permissions. As a result anyone might be able
to browse the state directory and read contacts/calendars.

Move to using fowners/fperms. Introduce a warning to tell users how to
fix it. Users that changed permissions on purpose can ignore the warning.

Signed-off-by: Henning Schild <henning@hennsch.de>