app-emulation/qemu: fix init.d/qemu-binfmt for mksh, bug #679168
Handling of escaped by echo is an implementation-defined behaviour
according to POSIX:
https://pubs.opengroup.org/onlinepubs/9699919799/utilities/echo.html

"A string to be written to standard output. If the first operand
is -n, or if any of the operands contain a <backslash> character,
the results are implementation-defined."

Practically this makes
    $ echo '\xFF'
behave differently on bash/dash and mksh: bash outputs '\xFF'
string of 4 bytes while mksh outputs single 0xFF into stdout.

The change is to use "printf '%s\n' '\xFF'" to get consistent
behaviour.

Reported-by: Jimmy.Jazz@gmx.net
Reviewed-by: Matthias Maier
Closes: https://bugs.gentoo.org/679168
Package-Manager: Portage-2.3.62, Repoman-2.3.12
Signed-off-by: Sergei Trofimovich <slyfox@gentoo.org>