net-p2p/syncthing: reset xdg environment
Turns out go uses $XDG_CACHE_HOME to decide where to put the build
cache. Make sure it, and to be on the safe side the rest of XDG
environment as well, is reset before building Syncthing.

We do not inherit xdg.eclass for this purpose because it introduces
several dependencies we do not actually need.

Closes: https://bugs.gentoo.org/679280
Signed-off-by: Marek Szuba <marecki@gentoo.org>
Package-Manager: Portage-2.3.62, Repoman-2.3.11