app-emulation/libvirt-9999: Update hash generation algorithm
I've noticed that bootstrap is run twice. This is because the
first time it is run in src_prepare() and there some hash is
computed so later when src_configre() gets its word it'll find
everything prepared. Well, in the upstream commit of
6d53255377b50b7e412350d079 the hash was changed therefore
src_configure() did not find the expected hash and thus re-ran
the bootstrap script.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Signed-off-by: Matthias Maier <tamiko@gentoo.org>