gnome-base/gnome-shell: restrict to older gjs
3.28 and older crash without upstream commit e36ba874
"Stop using conditional catch statements
It is a mozilla extension that is going away in SpiderMonkey 60."

It trivially applies to 3.28, after which 3.28 appears to work
fine with newer gjs, but we'll be bumping gnome-shell to 3.30
very soon anyways. Plus there seem to be non-crashing issues from
using newer gjs with gnome-shell-3.28 (missing labels in GPG/SSH
password dialogs for example).

Package-Manager: Portage-2.3.52, Repoman-2.3.12
Signed-off-by: Mart Raudsepp <leio@gentoo.org>