gnome-extra/cinnamon-session: Rely on IUSE=elogind for non-systemd
There are (probably obsolete) deps on polkit[elogind] for non-systemd
case, which is a problem now that we use.mask elogind on systemd
profiles. No-one hits this problem in practice (as elogind users
don't have systemd profiles active), but the CI and repoman trips up
on this.
Rework this into a REQUIRED_USE="^^ ( elogind systemd )" like in
other packages, until this polkit dep is there, to make repoman happy.

Package-Manager: Portage-2.3.52, Repoman-2.3.12
Signed-off-by: Mart Raudsepp <leio@gentoo.org>