profiles/targets/desktop/gnome: Enable USE=networkmanager by default
Networkmanager is a core component of GNOME, so default enable it.
Otherwise we just had a IUSE=+networkmanager in gnome-shell, which in
turn hits users with a need to package.use it for gnome-control-center
and more, or just globally enable it. Just do so by default, as it is
generally expected to be used with GNOME (yes, even with wired; easy
VPN support and so on).

Signed-off-by: Mart Raudsepp <leio@gentoo.org>