gnome-base/gdm: add elogind support
Include a patch from 3.32 that removes long gone legacy argument
to pam_systemd.so, as to make it uniform with the added elogind
support (and not add that legacy arg there too) and be able to
reuse the pam-elogind.patch bits for 3.32 without changes, just
dropping the argument patch.

The elogind support is achieved without configure.ac patches by
making use of the standard PKG_CHECK_MODULES builtin AC_ARG_VAR
provided variable override support and passing elogind values
there as a drop-in replacement. So we don't need an eautoreconf
and maintaing an autotools patch until upstream moves to meson.

Initial VT is put back to 7 for elogind for the time being until
there is more clarity what will happen with startDM.sh and how
things interact with agetty out of the box (potential keyboard
stealing, etc).

Closes: https://bugs.gentoo.org/645348
Package-Manager: Portage-2.3.52, Repoman-2.3.12
Signed-off-by: Mart Raudsepp <leio@gentoo.org>