gnome-base/gnome-extra-apps: drop default enabling of share temporarily
gnome-user-share currently requires systemd, which breaks clean
install of gnome metas out of the box for elogind systems.
Drop the IUSE default to be able to handle this at profile level
soon instead. This drops the default enable for systemd profiles
until we can add it back for systemd profiles (but we currently
lack a profiles/targets/desktop/gnome/systemd to easily do so)

Package-Manager: Portage-2.3.52, Repoman-2.3.12
Signed-off-by: Mart Raudsepp <leio@gentoo.org>