gnome-base/gnome-control-center: add elogind support
The power panel uses logind to check for suspend and hibernate
support. The color panel uses logind to disallow lid switch
and inhibit idle during calibration runs.

This revbump was accidentally not committed before. Should be
the last to enable "emerge gnome" to really work now on openrc
with elogind.

Consolekit option is removed, as the codebase actually doesn't
have any support for it whatsoever. openrc-settings dep remains
for the non-systemd path, as these services are used as well.

Closes: https://bugs.gentoo.org/681328
Package-Manager: Portage-2.3.52, Repoman-2.3.12
Signed-off-by: Mart Raudsepp <leio@gentoo.org>