sys-apps/util-linux: Build fix on MIPS from upstream
In glibc-2.29, the definitions of 'struct termios' were consolidated
across architectures.  This causes a build error in util-linux on MIPS
due to undefined struct members 'c_ospeed' and 'c_ispeed', which MIPS
does not have.  The included patch from upstream fixes this, and should
be in the next util-linux release.

Signed-off-by: Joshua Kinard <kumba@gentoo.org>
Package-Manager: Portage-2.3.62, Repoman-2.3.12