mail-filter/opendkim: add an elog about sharing a local socket.
Sharing a local OpenDKIM socket with an MTA is far too tricky. This
commit adds an elog with some vague hints, but we really need a decent
wiki page with some step-by-step instructions. This is progress
towards bug 575666, but I won't consider that resolved until we have
a wiki page.

Bug: https://bugs.gentoo.org/575666
Signed-off-by: Michael Orlitzky <mjo@gentoo.org>
Package-Manager: Portage-2.3.62, Repoman-2.3.11