mail-filter/opendkim: add a TODO question about a sed line.
There's a sed line in src_prepare that messes with some *.lua test
files. I've changed the action of that script to use "${T}" rather
than "/tmp" for security reasons... but I'm unable to figure out when
those test files are actuallty used. So, I've left a TODO comment to
figure it out in the future (or maybe just delete that sed line
entirely).

Signed-off-by: Michael Orlitzky <mjo@gentoo.org>
Package-Manager: Portage-2.3.62, Repoman-2.3.11