mail-filter/opendkim: new revision with a consistent config file.
The previous revisions installed a variable config file, and only did
so conditionally, based on the state of the live filesystem. The new
revision always installs the same config file. This should be simpler
and more predictable. In addition, the temporary directory "${T}" is
now used as "scratch space" while mangling that config file.

Signed-off-by: Michael Orlitzky <mjo@gentoo.org>
Package-Manager: Portage-2.3.62, Repoman-2.3.11