net-proxy/obfs4proxy: bump to 0.0.9
Partly vendored because trying to package both yawning/utls and
dsnet/compress resulted in them attempting to pull in a whole lot of
additional dependencies, which we apparently do not need in obfs4proxy
itself. Plus that way we were able to release this package without
waiting for dev-go/go-crypto, dev-go/go-net and dev-go/go-sys to be
bumped, although I hope the vendoring of these will only be temporary.

Closes: https://bugs.gentoo.org/677460
Signed-off-by: Marek Szuba <marecki@gentoo.org>
Package-Manager: Portage-2.3.62, Repoman-2.3.11