mail-filter/opendkim: Added support for a socket group
Installing this package will now create an additional group
"dkimsocket" which is meant to be shared between OpenDKIM and the
MTA. The group determines which processes have write access to the
local socket file. opendkim.conf now defaults to a local socket
instead of a TCP socket.

Signed-off-by: Ralph Seichter <gentoo@seichter.de>
Closes: https://bugs.gentoo.org/575666
Package-Manager: Portage-2.3.62, Repoman-2.3.11