app-text/gtranslator: bump to 3.32.0
Requires a recent python-based gtk-doc for some reason for API docs.
I suspect this is due to bugs in earlier python gtk-doc, while our perl
version would be fine.
As the gtk-docs are for internal API and not really necessary for most
people, just comment it out for now to get the new features and bug
fixes of the actual program.

Package-Manager: Portage-2.3.52, Repoman-2.3.12
Signed-off-by: Mart Raudsepp <leio@gentoo.org>