dev-util/gnome-builder-3.24: json-glib[introspection] is required for build
In 3.24 jsonrpc-glib is not a split out separate package, thus the global
introspection enabling on gnome-builder affects it as well (not optional).
jsonrpc-glib GIR requires json-glib GIR, thus add the missing introspection
USE dep on json-glib to ensure successful builds.
This is not a problem 3.26 and onwards, as jsonrpc-glib introspection
support is controlled by USE flag and there are proper conditional USE deps
in place in the external package on json-glib already there.

Fixes: https://bugs.gentoo.org/682060
Package-Manager: Portage-2.3.52, Repoman-2.3.12
Signed-off-by: Mart Raudsepp <leio@gentoo.org>