dev-python/sphinx: bump to 2.0.0
I removed the "net" USE flag because it does nothing else than adding
dependencies that aren't referenced anywhere in the code. Moreover, the
release that introduced it, 1.5.1, wasn't accompanied by a justification
for its addition. I think that it was added for what is now
"sphinxcontrib-websupport". If that's the case, then this doesn't belong
here.

Closes: https://bugs.gentoo.org/662322
Signed-off-by: Virgil Dupras <vdupras@gentoo.org>
Package-Manager: Portage-2.3.62, Repoman-2.3.11