mail-filter/opendkim: add "UMask 0117" to the config by default.
Ralph and I have been discussing the best way to handle a local UNIX
socket. One thing that we both agree on is that, regardless of the
other details, the daemon should be running with a UMask of 0117 to
create the socket group-writable but otherwise private. And since the
opendkim(8) man page says that the UMask is used only for the socket,
there's no reason we shouldn't add it to the default configuration. In
other words, it doesn't hurt anything if you're *not* using a local
socket. This is one fewer thing that we need to document.

Bug: https://bugs.gentoo.org/575666
Signed-off-by: Michael Orlitzky <mjo@gentoo.org>
Package-Manager: Portage-2.3.62, Repoman-2.3.11