net-mail/postfix-logwatch: maintainer stabilize -r3 on x86/amd64.
This package is pure perl, and so would have been marked ALLARCHES.
Moreover, it requires a special type of log file (that contains
sensitive information) to test. So, I think it's reasonable for me as
the maintainer to stabilize it on amd64, and (per ALLARCHES), on x86.

Signed-off-by: Michael Orlitzky <mjo@gentoo.org>
Package-Manager: Portage-2.3.62, Repoman-2.3.11