media-sound/playerctl: fix gdbus-codegen handling in configure.ac
In configure.ac, the **absolute** path of gdbus-codegen (as obtained
by pkg-config) is passed to AC_PATH_PROG. The macro then appends this
path to the directories in PATH and tests if the result is
executable. This internally results in checking for
'/usr/bin/usr/bin/gdbus-codegen'. The macro is redundant because the
value returned by pkg-config is sufficient.

Closes: https://bugs.gentoo.org/682788
Package-Manager: Portage-2.3.62, Repoman-2.3.12
Signed-off-by: Göktürk Yüksek <gokturk@gentoo.org>