dev-python/pytest-cov: relax pytest-xdist dep constraint
Tests were already failing badly before, for unrelated reasons and this
is the only package holding pytest-xdist back (I don't even think that
it's actually needed for testing...).

Also, make the Python project the maintainer of this package. chutzpah
doesn't have a single commit for this package in the git log. Everybody
but him touched that package.

Signed-off-by: Virgil Dupras <vdupras@gentoo.org>
Package-Manager: Portage-2.3.62, Repoman-2.3.11