sys-fs/lvm2: don't get stuck with root-on-LVM
During system shutdown, init.d/lvm tries to run 'vgchange -a ln' to
close all LVs. It fails with root-on-LVM (or seperate /usr-on-LVM),
because those mountpoints are still in use. As a best-effort, consider
it done after the 'vgchange -a ln' pass. This is similar to localmount's
stop, which does not fail if a filesystem cannot be unmounted.

Package-Manager: Portage-2.3.62, Repoman-2.3.12
Signed-off-by: Robin H. Johnson <robbat2@gentoo.org>