app-crypt/gnupg: drop virtual/mta requirement due to stage3 impact
In the previous fix for bug 648606, virtual/mta was made a hard DEPEND
of gnupg. This has the side-effect of adding the default virtual/mta
(nullmailer) and the default virtual/logger (app-admin/metalog) to all
stage3 builds.

As a better fix, ensure the path to the mailprog is hardcoded and does
not depend on the presence of the virtual/mta package, using
/usr/libexec/sendmail, as that's what is provided by all virtual/mta
providers.

Thereafter, drop the direct virtual/mta dependency and add a
pkg_postinst describing the tiny piece of functionality that depends on
the presence of virtual/mta: 'gpg-wks-client --send'
'gpg-wks-server' also uses virtual/mta, but is already behind a USE
flag, so ensure it's present there instead.

See-Also: https://bugs.gentoo.org/648606
Package-Manager: Portage-2.3.62, Repoman-2.3.12
Signed-off-by: Robin H. Johnson <robbat2@gentoo.org>