dev-lang/go: add system-bootstrap use flag
If a previous version of dev-lang/go is installed, you can use this use
flag to skip downloading the bootstrap archives and bootstrap the new
version using the already-installed version.

For the initial installation of dev-lang/go, we have to download all of
the bootstrap archives to make the go ebuild work under crossdev. For
this reason, system-bootstrap cannot be turned on by default.

Bug: https://bugs.gentoo.org/671394
closes: https://bugs.gentoo.org/684652
Copyright: Sony Interactive Entertainment Inc.
Package-Manager: Portage-2.3.62, Repoman-2.3.12
Signed-off-by: William Hubbs <williamh@gentoo.org>