llvm.eclass: Add EAPI 7 API to get_llvm_prefix
Add two switches to get_llvm_prefix(), '-b' and '-d' to enable use
of LLVM API cross support.  '-b' is intended to be used whenever prefix
to CBUILD llvm-config is needed, and '-d' (the default) is intended
to be used whenever prefix to CHOST bindir is needed (to find CMake
modules).

Signed-off-by: Michał Górny <mgorny@gentoo.org>