app-emulation/docker-credential-helpers: new package
New ebuild for docker-credential-helpers, which has helper programs for
docker to store your password and other sensitive information.

This ebuild builds either docker-credential-secretservice, which uses
the D-Bus secret service API to talk to either kwallet or gnome-keyring;
or docker-credential-pass which relies on the pass utility; or both
depending on USE flags.

The supported USE flags are:

- 'gnome-keyring': build secretservice helper

- 'pass': build pass helper

A message is printed via elog() upon successful merge about how to use
the helper with docker.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>
Closes: https://github.com/gentoo/gentoo/pull/9805
Signed-off-by: Michał Górny <mgorny@gentoo.org>