sys-libs/db: compile-fix by dropping mutex=ARM/gcc-assembly
This is a compile-fix so that old versions build on newer ARM hardware.

The ARM assembly code does not work on newer hardware
so you CANNOT use --with-mutex=ARM/gcc-assembly anymore.

Specifically, it uses the SWPB op, which was deprecated [1]
The SWPB op ALSO cannot be used in ARM-Thumb mode.

>=sys-libs/db-6.1 uses LDREX instead, but that fix CANNOT be backported
due differing licenses in the new versions (the new version is AGPL).

Trust the compiler instead for older sys-libs/db versions, and just drop
--with-mutex=ARM/gcc-assembly from the configure.

[1] http://www.keil.com/support/man/docs/armasm/armasm_dom1361289909499.htm

Fixes: https://bugs.gentoo.org/555740
Reported-By: Christohper Harrington <ironiridis@gmail.com>
Reported-By: Luke-Jr <luke-jr@utopios.org>
Package-Manager: Portage-2.3.62, Repoman-2.3.12
Signed-off-by: Robin H. Johnson <robbat2@gentoo.org>