app-text/mupdf: fix opengl dependencies
It was a very bad idea from me to remove dependency on freeglut
on the account that we use the bundled version. I completely forgot
about freeglut dependencies themselves. Bad me.

These dependencies, however, are a bit complex and I find it fragile to
copy-paste them in mupdf. At the cost of installing an extra system
library (freeglut) that is rather small and probably already installed
anyways, we potentially save ourselves tricky problems down the road.

Closes: https://bugs.gentoo.org/685604
Signed-off-by: Virgil Dupras <vdupras@gentoo.org>
Package-Manager: Portage-2.3.62, Repoman-2.3.11