net-wireless/iwd: bump to 0.18 & add regulatory domain checks
Added new default-on dep on crda.  One can disable this flag & dep, to
depend on the new in-kernel support, if so desired.  This requires the
cfg80211 module to load firmware files when it initializes, so the
usual firmware-loading requirements apply.

I am not currently checking for the regulatory.db.p7s file that's
required if CFG80211_REQUIRE_SIGNED_REGDB=y is set, that will be
improved in the future.

Closes: https://bugs.gentoo.org/683060
Closes: https://bugs.gentoo.org/685424
Closes: https://bugs.gentoo.org/685588
Package-Manager: Portage-2.3.66, Repoman-2.3.12
RepoMan-Options: --force
Signed-off-by: Ben Kohler <bkohler@gentoo.org>