games-strategy/openra: Bump to 20190314 and major overhaul
Thanks to Yaroslav Yarigin for his ebuild, which I based this on.

Many additional DLLs are needed, which are normally downloaded during
the build but I have added them to SRC_URI.

I am ditching the Gentoo README because although a lot of effort had
been put into it, the game content is now dead simple to install and
the upstream documentation is first class so it's really redundant.

Closes: https://bugs.gentoo.org/582942
Package-Manager: Portage-2.3.66, Repoman-2.3.12
Signed-off-by: James Le Cuirot <chewi@gentoo.org>